silence select() first arg warnings with casts on Windows
It's an unused arg on that platform.

```
example/direct_tcpip.c:264:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/direct_tcpip.c:264:33: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/direct_tcpip.c:264:33: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/scp_nonblock.c:81:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/scp_nonblock.c:81:27: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/scp_nonblock.c:81:27: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/scp_write_nonblock.c:68:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/scp_write_nonblock.c:68:27: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/scp_write_nonblock.c:68:27: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/sftp_RW_nonblock.c:75:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/sftp_RW_nonblock.c:75:27: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/sftp_RW_nonblock.c:75:27: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/sftp_RW_nonblock.c:268:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/sftp_RW_nonblock.c:268:26: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/sftp_RW_nonblock.c:268:26: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/sftp_RW_nonblock.c:328:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/sftp_RW_nonblock.c:328:30: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/sftp_RW_nonblock.c:328:30: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/sftp_nonblock.c:86:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/sftp_nonblock.c:86:27: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/sftp_nonblock.c:86:27: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/sftp_write_nonblock.c:74:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/sftp_write_nonblock.c:74:27: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/sftp_write_nonblock.c:74:27: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/sftp_write_sliding.c:74:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/sftp_write_sliding.c:74:27: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/sftp_write_sliding.c:74:27: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/ssh2_agent_forwarding.c:79:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/ssh2_agent_forwarding.c:79:27: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/ssh2_agent_forwarding.c:79:27: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/ssh2_echo.c:75:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/ssh2_echo.c:75:27: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/ssh2_echo.c:75:27: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/ssh2_exec.c:77:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/ssh2_exec.c:77:27: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/ssh2_exec.c:77:27: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
example/tcpip-forward.c:254:..: warning C4244: 'function': conversion from 'libssh2_socket_t' to 'int', possible loss of data
example/tcpip-forward.c:254:33: warning: conversion from 'libssh2_socket_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
example/tcpip-forward.c:254:33: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
```

+++ example/scp_write_nonblock.c
@@ -65,7 +65,7 @@ static int waitsocket(libssh2_socket_t socket_fd, LIBSSH2_SESSION *session)
     if(dir & LIBSSH2_SESSION_BLOCK_OUTBOUND)
         writefd = &fd;
 
-    rc = select(socket_fd + 1, readfd, writefd, NULL, &timeout);
+    rc = select((int)(socket_fd + 1), readfd, writefd, NULL, &timeout);
 
     return rc;
 }
